Custom Auth

+++ app/Http/Controllers/PostController.php
@@ -7,13 +7,14 @@ use App\Models\Post;
 use App\Models\Tag;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class PostController extends Controller
 {
     public function getIndex()
     {
         // $posts = Post::orderBy('created_at','desc')->get();
-        $posts = Post::orderBy('created_at', 'desc')->paginate(3);
+        $posts = Post::orderBy('created_at', 'desc')->paginate(2);
         return view('blog.index', ['posts' => $posts]);
     }
 
@@ -34,18 +35,27 @@ class PostController extends Controller
 
     public function getAdminIndex()
     {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
         $posts = Post::orderBy('created_at', 'desc')->get();
         return view('admin.index', ['posts' => $posts]);
     }
 
     public function getAdminCreate()
     {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
         $tags = Tag::all();
         return view('admin.create', ['tags' => $tags]);
     }
 
     public function getAdminEdit($id)
     {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
         $post = Post::find($id);
         $tags = Tag::all();
         return view('admin.edit', ['post' => $post, 'tags' => $tags]);
@@ -53,6 +63,10 @@ class PostController extends Controller
 
     public function postAdminCreate(Request $request)
     {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
+
         $this->validate($request, [
             'title' => 'required|min:5',
             'content' => 'required|min:10',
@@ -62,11 +76,12 @@ class PostController extends Controller
         if (!$user) {
             return redirect()->back();
         }
+
         $post = new Post([
             'title' => $request->input('title'),
             'content' => $request->input('content'),
         ]);
-//        $post->save();
+        //$post->save();
         $user->posts()->save($post);
         $post->tags()->attach($request->input('tags') === null ? [] : $request->input('tags'));
 
@@ -76,16 +91,23 @@ class PostController extends Controller
 
     public function postAdminUpdate(Request $request)
     {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
+
         $this->validate($request, [
             'title' => 'required|min:5',
             'content' => 'required|min:10',
         ]);
         $post = Post::find($request->input('id'));
+        if (Gate::denies('update-post', $post)) {
+            return redirect()->route('admin.index')->with('info', 'Your are not authorised!');
+        }
         $post->title = $request->input('title');
         $post->content = $request->input('content');
         $post->save();
-        //        $post->detach();
-        //        $post->tags()->attach($request->input('tags') === null ? [] : $request->input('tags'));
+        //$post->detach();
+        //$post->tags()->attach($request->input('tags') === null ? [] : $request->input('tags'));
         $post->tags()->sync($request->input('tags') === null ? [] : $request->input('tags'));
 
         return redirect()->route('admin.index')
@@ -94,7 +116,14 @@ class PostController extends Controller
 
     public function getAdminDelete($id)
     {
+        if (!Auth::check()) {
+            return redirect()->back();
+        }
+
         $post = Post::find($id);
+        if (Gate::denies('update-post', $post)) {
+            return redirect()->back()->with('info', 'Your are not authorised!');
+        }
         $post->likes()->delete();
         $post->tags()->detach();
         $post->delete();
